Combine some files
