--- SuperpositionGame/src/main/scala/superposition/math/DependentKey.scala
@@ -1,5 +1,0 @@
-package superposition.math
-
-trait DependentKey {
-  type Value
-}
