--- SuperpositionGame/src/main/scala/superposition/math/DependentMap.scala
@@ -1,3 +1,0 @@
-package superposition.math
-
-trait DependentMap[K <: DependentKey] extends DependentMapOps[K, DependentMap[K]]
