--- SuperpositionGame/src/main/scala/superposition/quantum/MetaId.scala
@@ -1,7 +1,0 @@
-package superposition.quantum
-
-import superposition.math.DependentKey
-
-final class MetaId[A] extends DependentKey {
-  type Value = A
-}
